requirements.txt
# Core ETL
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0

# PostgreSQL
sqlalchemy>=2.0
psycopg2-binary>=2.9

# Dashboard Director
streamlit>=1.35
plotly>=5.20

# Reportes Excel
xlsxwriter>=3.1

# QA
black>=24.0
flake8>=7.0
mypy>=1.10
pytest>=8.0
